Add performance tests for Client commands with Cosmos and Postgres integration
